+++ src/main/java/br/ufpb/dcx/apps4society/educandoapi/domain/Context.java
@@ -27,7 +27,7 @@ public class Context implements Serializable {
 	private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String name;
 	@JsonIgnore
 	@ManyToOne
@@ -58,7 +58,7 @@ public class Context implements Serializable {
 	 * @param sound The sound for this Context.
 	 * @param videoUrl The URL of a video for this Context.
 	 */
-	public Context(Integer id, String name,  User creator, String imageUrl, String soundUrl, String videoUrl) {
+	public Context(Long id, String name,  User creator, String imageUrl, String soundUrl, String videoUrl) {
 		this.id = id;
 		this.name = name;
 		this.creator = creator;
@@ -91,7 +91,7 @@ public class Context implements Serializable {
 	 * 
 	 * @return the if of this Context.
 	 */
-	public Integer getId() {
+	public Long getId() {
 		return this.id;
 	}
 
@@ -101,7 +101,7 @@ public class Context implements Serializable {
 	 * @param id
 	 *            The new value of the id for this Context.
 	 */
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
